Add NotificationProvider using IFTTT.com
Using the IFTTT Maker-channel and from there it can trigger any number
of systems. I've connected mine to my phone and my Philips HUE lamps.
The Maker-channel is for devs mostly, so anyone that wants notifications
needs to sign up and send in their key to the demo.

Use INotificationPreferenceRegister to lookup Maker-keys. Hardcoded to
keep things simple.